+++ CHANGES.md
@@ -7,8 +7,13 @@ _Note:_ pyfakefs 4.0 is planned for release at the end of 2019 or the beginning
 of 2020. As pyfakefs 4.0 is a major release, we are giving you advance notice of
 the proposed changes so you can be ready.
 
-  * pyfakefs 4.0 drops support for Python 2.7. If you still need
-    Python 2.7, you can continue to use the latest pyfakefs 3.x version. 
+## [Version 3.7.2](https://pypi.python.org/pypi/pyfakefs/3.7.2)
+
+This version backports some fixes from master.
+
+#### Fixes
+  * Fixed handling of `os.TMPFILE` flag under Linux
+    (see [#509](../../issues/509))
 
 ## [Version 3.7.1](https://pypi.python.org/pypi/pyfakefs/3.7.1)